pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "homeview"
version = "0.1.0"
description = "Self-hosted multi-stream video wall controller"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "aiosqlite>=0.20.0",
    "python-xlib>=0.33",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
]

[project.scripts]
homeview = "server.cli:start"
homeview-reset-pairing = "server.cli:reset_pairing"

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "anyio[trio]>=4.6.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["server"]

[dependency-groups]
dev = [
    "pytest-cov>=7.1.0",
]
